added new features and fixed #responsive bugs
